Refactor Chart into its own component
